Moved ComponentForPathConfig
